Update readme.md
Should we add a 4th roll?  That would be a tech vendor that actually does the tokenization.  It seems that project owners, VVB and Registries will not be capable of gathering DIDs, normalizing the data and actually tokenizing the asset.  Not sure this should be a formal role, but maybe.